Move `DiskFarm` to `shared` module
