+++ crates/subspace-farmer/src/bin/subspace-farmer/commands/farm.rs
@@ -1,5 +1,6 @@
 use crate::commands::shared::metrics::{FarmerMetrics, SectorState};
 use crate::commands::shared::network::{configure_network, NetworkArgs};
+use crate::commands::shared::DiskFarm;
 use crate::utils::shutdown_signal;
 use anyhow::anyhow;
 use async_lock::Mutex as AsyncMutex;
@@ -13,7 +14,6 @@ use parking_lot::Mutex;
 use prometheus_client::registry::Registry;
 use std::net::SocketAddr;
 use std::num::{NonZeroU8, NonZeroUsize};
-use std::path::PathBuf;
 use std::pin::pin;
 use std::str::FromStr;
 use std::sync::Arc;
@@ -270,68 +270,6 @@ fn cache_percentage_parser(s: &str) -> anyhow::Result<NonZeroU8> {
     }
 
     Ok(cache_percentage)
-}
-
-#[derive(Debug, Clone)]
-pub(crate) struct DiskFarm {
-    /// Path to directory where data is stored.
-    directory: PathBuf,
-    /// How much space in bytes can farm use for plots (metadata space is not included)
-    allocated_plotting_space: u64,
-}
-
-impl FromStr for DiskFarm {
-    type Err = String;
-
-    fn from_str(s: &str) -> anyhow::Result<Self, Self::Err> {
-        let parts = s.split(',').collect::<Vec<_>>();
-        if parts.len() != 2 {
-            return Err("Must contain 2 coma-separated components".to_string());
-        }
-
-        let mut plot_directory = None;
-        let mut allocated_plotting_space = None;
-
-        for part in parts {
-            let part = part.splitn(2, '=').collect::<Vec<_>>();
-            if part.len() != 2 {
-                return Err("Each component must contain = separating key from value".to_string());
-            }
-
-            let key = *part.first().expect("Length checked above; qed");
-            let value = *part.get(1).expect("Length checked above; qed");
-
-            match key {
-                "path" => {
-                    plot_directory.replace(PathBuf::from(value));
-                }
-                "size" => {
-                    allocated_plotting_space.replace(
-                        value
-                            .parse::<ByteSize>()
-                            .map_err(|error| {
-                                format!("Failed to parse `size` \"{value}\": {error}")
-                            })?
-                            .as_u64(),
-                    );
-                }
-                key => {
-                    return Err(format!(
-                        "Key \"{key}\" is not supported, only `path` or `size`"
-                    ));
-                }
-            }
-        }
-
-        Ok(DiskFarm {
-            directory: plot_directory.ok_or({
-                "`path` key is required with path to directory where plots will be stored"
-            })?,
-            allocated_plotting_space: allocated_plotting_space.ok_or({
-                "`size` key is required with path to directory where plots will be stored"
-            })?,
-        })
-    }
 }
 
 /// Start farming by using multiple replica plot in specified path and connecting to WebSocket
