+++ crates/subspace-farmer/src/bin/subspace-farmer/commands/shared.rs
@@ -1,8 +1,72 @@
 pub(super) mod metrics;
 pub(super) mod network;
 
+use bytesize::ByteSize;
 use std::path::PathBuf;
+use std::str::FromStr;
 use subspace_farmer::single_disk_farm::{SingleDiskFarm, SingleDiskFarmSummary};
+
+#[derive(Debug, Clone)]
+pub(in super::super) struct DiskFarm {
+    /// Path to directory where data is stored.
+    pub(in super::super) directory: PathBuf,
+    /// How much space in bytes can farm use for plots (metadata space is not included)
+    pub(in super::super) allocated_plotting_space: u64,
+}
+
+impl FromStr for DiskFarm {
+    type Err = String;
+
+    fn from_str(s: &str) -> anyhow::Result<Self, Self::Err> {
+        let parts = s.split(',').collect::<Vec<_>>();
+        if parts.len() != 2 {
+            return Err("Must contain 2 coma-separated components".to_string());
+        }
+
+        let mut plot_directory = None;
+        let mut allocated_plotting_space = None;
+
+        for part in parts {
+            let part = part.splitn(2, '=').collect::<Vec<_>>();
+            if part.len() != 2 {
+                return Err("Each component must contain = separating key from value".to_string());
+            }
+
+            let key = *part.first().expect("Length checked above; qed");
+            let value = *part.get(1).expect("Length checked above; qed");
+
+            match key {
+                "path" => {
+                    plot_directory.replace(PathBuf::from(value));
+                }
+                "size" => {
+                    allocated_plotting_space.replace(
+                        value
+                            .parse::<ByteSize>()
+                            .map_err(|error| {
+                                format!("Failed to parse `size` \"{value}\": {error}")
+                            })?
+                            .as_u64(),
+                    );
+                }
+                key => {
+                    return Err(format!(
+                        "Key \"{key}\" is not supported, only `path` or `size`"
+                    ));
+                }
+            }
+        }
+
+        Ok(DiskFarm {
+            directory: plot_directory.ok_or({
+                "`path` key is required with path to directory where plots will be stored"
+            })?,
+            allocated_plotting_space: allocated_plotting_space.ok_or({
+                "`size` key is required with path to directory where plots will be stored"
+            })?,
+        })
+    }
+}
 
 pub(super) fn print_disk_farm_info(directory: PathBuf, farm_index: usize) {
     println!("Single disk farm {farm_index}:");
